fix(contracts): update `L2OutputOracle` constructor require statement
Updates the constructor check in the `L2OutputOracle` so that it is correct.
It previously compared two different types of units (blocknumbers and
timestamps). The real thing that we need to ensure is that neither is 0 because
that will result in a broken system.

See https://github.com/ethereum-optimism/optimism/pull/5478