revert!: remove blue hard fork and version L2 output root to 0
Since we determined to start with new chain ID, not hard fork, removed the codes
related to blue hard fork.
BREAKING CHANGE: Previously we set the L2 output root to version 1 after the
blue hard fork, but we decided not to hard fork, dropped that version to 0. The
output root will vary accordingly.

Refs: 6d82245, 62cfaef, 83e3f2b