fix(contracts): immediately fail when re-entry `relayMessage`
`CrossDomainMessager` whereby `relayMessage` naively reset the
`xDomainMsgSender` after a call. The issue here was that a nested call would
reset this value to the `Constants.DEFAULT_L2_SENDER` which deviates from the
expected behavior that the `xDomainMsgSender` is set to the `_sender` for the
entirety of the subcall.
This change sets a message as failed upon re-entry to the `relayMessage`
function by checking the value of `xDomainMsgSender`. We then also remove the
`reentrancyLocks` from storage and don't have to worry about re-entrency here
since any subcall through the `SafeCall` lib to the XDM `relayMessage` will be
caught and recorded as failed by the aforementioned `xDomainMsgSender` check.
Also removes the revert in `relayMessage` when the minimum gas threshold is not
met.

See:
  - https://github.com/ethereum-optimism/optimism/pull/5475
  - https://github.com/ethereum-optimism/optimism/pull/5444
  - https://github.com/ethereum-optimism/optimism/pull/5493
  - https://github.com/ethereum-optimism/optimism/pull/5508