feat(batcher): properly drain state on L2 reorgs
When an L2 reorg occurs the batcher should close out the pending channel and
submit everything that it has before clearing the channel builder. The reason
for this is that even though it may be submitting incorrect data it should not
leave dangling channels. After the pending state is drained it can begin
tracking the new L2 chain.

See Optimism PR#5536.