fix(contracts): update `SystemConfig` resource config validation
Updates the way that the system config validates the resource config to ensure
that it's not possible to set config values that break the system. It is
possible to set config values that would burn too much gas, rendering the system
unusable. Every error cannot be caught, so try to catch the most possible.
Altering the default resource config is not recommended, but it is possible.

This also adds fuzz tests that cover the problem, if the change in the require
statement in the system config contract is reverted, then the fuzz tests catch
the `UNDEFINED` error quite quickly. For this test, increases
`startingBlockNumber` to prevent underflow error.

Also adds a `ResourceMetering` test to demonstrate that resource metering fails
if the `baseFeeMaxChangeDenominator` is set to 1.

See Optimism PR#5445, PR#5499.