fix(validator): prevent challenger restarting while sync blocks
Challenger continuously restarts when kroma node is syncing. To prevent it,
changes to loop when error occurred, not restart. And check sync status in
validator initially, because currently all the components in validator
commonly check it. We do not log sync status error while syncing because node
always does not respond while syncing.