feat(contracts): change to select the next validator when output submission
Current, the next priority validator is selected when unbond based on the last
finalized output. However, if the output of `nextUnbondOutputIndex` is submitted
in public round, no output will be unbonded while `tryUnbond()`,
leading to the same validator to be kept.
To select the next validator for every submission randomly,
the next priority validator is selected when output submission.