fix(contracts): fix some bugs related to `AssetManager` (#327)
* fix(contracts): correctly subtract from validator delegated KRO when undelegate

When undelegate, only validator delegated KRO should be subtracted from
`validatorKro`. The accumulated base reward should not be subtracted from
`validatorKro`.

* fix(contracts): correct the input of `_convertToKroAssets`

Previously, the input of `_convertToKroAssets` was wrong, so correct it from KGH
shares to token ID. Also organize some public functions of `AssetManager` since
previous version of `previewKghUndelegate` is not appropriate for public use.

* fix(contracts): replace `msg.sender` in view function with argument

Since `msg.sender` in view function is zero address, we should not use it.

* feat(contracts): add getter of `validatorRewardKro`

* chore: rename `assetMan` to `assetMgr`

* feat(contracts): don't need to update validator tree before jail

Since the validator will be removed from the validator tree when jailed, we
don't need to update validator tree before sending the validator to the jail.

* fix(contracts): update validator KRO when slashing

`validatorKro` is included in `totalKro`, so it does not need to be included in
calculation of slashing amount. But we need to update this value as well when
modifying each asset of vault when applying slash.

* chore(contracts): update bindings and snapshots