fix(validator): change challenger and guardian scan until head L1 (#330)
When using `syncStatus.CurrentL1` instead of `syncStatus.HeadL1`, it is
impossible to scan between `CurrentL1` and `HeadL1`. If an asserter was off when
a challenge was created, the asserter would not be able to react when turned on
right after the event occurred. Guardian could also have the same issue, so
fixed it also.