feat(contracts): replace output submitter to SC in `dismissChallenge`
When a malicious challenge successfully proved fault due to a ZK bug, the
asserter will be slashed and not have bond anymore. But after challenge
dismissed by security council, the output submitter is reverted to the asserter
and the asserter's asset will be tried to be unbonded when output being
finalized. In that case, the asserter does not have bond since it has been
slashed, so output finalization will be reverted.

I made a change to replace the output submitter to security council when
dismissing challenge, and block proving fault when submitter is security
council because security council has already validated the output.